code refactoring
