+++ src/index.js
@@ -2,10 +2,10 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-import App from "./App";
-import "./index.css";
 import reportWebVitals from "./reportWebVitals";
-import Layout from "./components/Layout";
+import "./index.css";
+import App from "./App";
+import Draw from "./pages/Draw";
 
 const router = createBrowserRouter([
   {
@@ -14,7 +14,7 @@ const router = createBrowserRouter([
   },
   {
     path: "/draw",
-    element: <Layout />,
+    element: <Draw />,
   },
 ]);
